Crate futures_borrow

Futures-aware borrow cell.

Given the asynchronous nature of working with futures, managing borrows that live across callbacks can be difficult. This is because lifetimes cannot be moved into closures that are 'static (i.e. most of the futures-rs combinators).

Borrow provides runtime checked borrowing, similar to RefCell, however Borrow also provides Future task notifications when borrows are dropped.

Structs

Borrow
A mutable memory location with future-aware dynamically checked borrow rules.

BorrowError
Error produced by a failed poll_borrow call.

BorrowGuard
Holds a borrowed value obtained from Borrow.

TryBorrowError
Error produced by a failed try_borrow call.